requirements.txt
google-genai==0.2.1
google-generativeai==0.8.3
streamlit==1.41.0